Added getAllIDs like getUserID but not limited to users(depending on arg), re-wrote getUserID to call it
