+++ README.md
@@ -59,6 +59,7 @@ Result:
 * [`api.getThreadInfo`](DOCS.md#getThreadInfo)
 * [`api.getThreadList`](DOCS.md#getThreadList)
 * [`api.getThreadPictures`](DOCS.md#getThreadPictures)
+* [`api.getAllIDs`](DOCS.md#getAllIDs)
 * [`api.getUserID`](DOCS.md#getUserID)
 * [`api.getUserInfo`](DOCS.md#getUserInfo)
 * [`api.handleMessageRequest`](DOCS.md#handleMessageRequest)
