+++ src/getUserID.js
@@ -1,6 +1,6 @@
 "use strict";
 
-var utils = require("../utils");
+var all_ids = require("./getAllIDs.js");
 var log = require("npmlog");
 
 function formatData(data) {
@@ -17,39 +17,8 @@ function formatData(data) {
 }
 
 module.exports = function(defaultFuncs, api, ctx) {
+  var getAllIDs = all_ids(defaultFuncs,api,ctx);
   return function getUserID(name, callback) {
-    if(!callback) {
-      throw {error: "getUserID: need callback"};
-    }
-
-    var form = {
-      'value' : name.toLowerCase(),
-      'viewer' : ctx.userID,
-      'rsp' : "search",
-      'context' : "search",
-      'path' : "/home.php",
-      'request_id' : utils.getGUID(),
-    };
-
-    defaultFuncs
-      .get("https://www.facebook.com/ajax/typeahead/search.php", ctx.jar, form)
-      .then(utils.parseAndCheckLogin(ctx, defaultFuncs))
-      .then(function(resData) {
-        if (resData.error) {
-          throw resData;
-        }
-
-        var data = resData.payload.entries;
-
-        if(data[0].type !== "user") {
-          throw {error: "Couldn't find a user with name " + name + ". Bes match: " + data[0].path};
-        }
-
-        callback(null, data.map(formatData));
-      })
-      .catch(function(err) {
-        log.error("getUserID", err);
-        return callback(err);
-      });
+    return getAllIDs(name,false,callback);
   };
 };
